+++ FloripaPublicTransportation/app/src/main/java/com/arctouch/floripapublictransportation/activity/DetailsRouteActivity.java
@@ -1,9 +1,11 @@
-package com.arctouch.floripapublictransportation;
+package com.arctouch.floripapublictransportation.activity;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class DetailsActivity extends AppCompatActivity {
+import com.arctouch.floripapublictransportation.R;
+
+public class DetailsRouteActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
