+++ FloripaPublicTransportation/app/src/main/res/raw/configuration.properties
@@ -1,0 +1,5 @@
+user = WKD4N7YMA1uiM8V
+password = DtdTtzMLQlA0hk2C1Yi5pLyVIlAQ68
+urlFindRoutes = https://api.appglu.com/v1/queries/findRoutesByStopName/run
+urlFindStops = https://api.appglu.com/v1/queries/findStopsByRouteId/run
+urlFindDepartures = https://api.appglu.com/v1/queries/findDeparturesByRouteId/run
